requirements.txt
backports.ssl-match-hostname==3.4.0.2
certifi==14.5.14
enum34==1.0.4
factory-boy==2.4.1
fake-factory==0.4.2
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
pycountry==1.10
SQLAlchemy==0.9.8
tornado==4.0.2
torndb==0.3
